+++ src/mesh/mesh_group.hpp
@@ -13,15 +13,15 @@ struct MeshGroup
 public:
     // Define class attributes
     Mesh**          meshes              = nullptr;
-    size_t          meshes_np           = 0;
+    int             meshes_np           = 0;
     PanelGeom**     panels              = nullptr;
-    size_t*         panels_np           = nullptr;
-    size_t*         panels_cnp          = nullptr;
-    size_t          panels_tnp          = 0;
+    int*            panels_np           = nullptr;
+    int*            panels_cnp          = nullptr;
+    int             panels_tnp          = 0;
     SourceNode**    source_nodes        = nullptr;
-    size_t*         source_nodes_np     = nullptr;
-    size_t*         source_nodes_cnp    = nullptr;
-    size_t          source_nodes_tnp    = 0;
+    int*            source_nodes_np     = nullptr;
+    int*            source_nodes_cnp    = nullptr;
+    int             source_nodes_tnp    = 0;
 
     // Define class constructor and destructor
     MeshGroup(
